+++ CHANGELOG.md
@@ -15,7 +15,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
-* Fixed the bug when function `convertIfDate` converted a string with a date inside into a date object `core/json`
+* Fixed a bug when the function `convertIfDate` converted a string with a date inside into a date object `core/json`
 
 ## v3.58.1 (2021-09-06)